backend/requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
flask-restx==1.3.0
PyMySQL==1.1.1
python-dotenv==1.1.1
Flask-CORS==6.0.1
Flask-JWT-Extended==4.7.1